requirements.txt
allure-pytest~=2.13.5
selene==2.0.0rc9
selenium~=4.27.1
pytest==8.3.3
requests==2.32.3


allure-python-commons~=2.13.5